Rename testcase.TestCase to test.Case
